requirements.txt
opencv-python
numpy
Pillow
tk